Rename src/components/dashboard to src/components/home.
