+++ src/components/home/metrics/NetworkMetricsLoader.ts
